chore: Split work files from home page
