+++ react-redux-ts/src/components/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { Work } from "./Work";
